+++ src/main/scala/edu/ie3/simona/model/participant/PVModel.scala
@@ -740,8 +740,7 @@ final case class PVModel private (
       .multiply(genCorr * tempCorr)
 
     /* Calculate the foreseen active power output without boundary condition adaptions */
-    val proposal = sRated
-      .multiply(-1)
+    val activePower = sRated
       .multiply(
         `yield`
           .divide(irraditionSTC)
@@ -752,19 +751,12 @@ final case class PVModel private (
       .asType(classOf[Power])
       .to(MEGAWATT) // MW.
 
-    /* Do sanity check, if the proposed feed in is above the estimated maximum to be apparent active power of the plant */
-    if (proposal.isLessThan(pMax))
-      logger.warn(
-        "The fed in active power is higher than the estimated maximum active power of this plant ({} < {}). " +
-          "Did you provide wrong weather input data?",
-        proposal,
-        pMax
-      )
+    /* If the output is marginally small, suppress the output, as it's likely nighttime where there shouldn't be any output */
+    if (activePower.compareTo(activationThreshold) > 0)
+      return Quantities.getQuantity(0d, MEGAWATT)
 
-    /* If the output is marginally small, suppress the output, as we are likely to be in night and then only produce incorrect output */
-    if (proposal.compareTo(activationThreshold) > 0)
-      Quantities.getQuantity(0d, MEGAWATT)
-    else proposal
+    /* Do sanity check, if the proposed feed in is above the estimated maximum active power of the plant */
+    limitActivePower(activePower, pMax).multiply(-1)
   }
 }