+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement SQL source for primary data [#34](https://github.com/ie3-institute/simona/issues/34), [#101](https://github.com/ie3-institute/simona/issues/101)
 
 ### Changed
+- Harmonized active power output of `WecInput` and `PvInput` [#135](https://github.com/ie3-institute/simona/issues/135)
 - Improving code readability in EvcsAgent by moving FreeLotsRequest to separate methods
 - Re-organizing test resources into their respective packages [#105](https://github.com/ie3-institute/simona/issues/105)
 - BREAKING: Using snapshot version of PSDM